Rename kh.tools.layout to OpenKh.Tools.LayoutViewer
