+++ OpenKh.Tools.Common/CommandHelpers.cs
