+++ OpenKh.Tools.Common/Controls/DrawPanel.cs
