+++ OpenKh.Tools.Common/DependencyPropertyUtils.cs
