+++ OpenKh.Tools.Common/Utilities.cs
