+++ OpenKh.Tools.LayoutViewer/App.config
