+++ OpenKh.Tools.LayoutViewer/App.xaml
@@ -1,7 +1,7 @@
-﻿<Application x:Class="kh.tools.layout.App"
+﻿<Application x:Class="OpenKh.Tools.LayoutViewer.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:kh.tools.layout"
+             xmlns:local="clr-namespace:OpenKh.Tools.LayoutViewer"
              StartupUri="Views/MainWindow.xaml">
     <Application.Resources>
