+++ OpenKh.Tools.LayoutViewer/Properties/Resources.resx
