+++ OpenKh.Tools.LayoutViewer/Properties/Settings.settings
