+++ OpenKh.Tools.LayoutViewer/Views/FrameView.xaml
@@ -1,9 +1,9 @@
-﻿<UserControl x:Class="kh.tools.layout.Views.FrameView"
+﻿<UserControl x:Class="OpenKh.Tools.LayoutViewer.Views.FrameView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:kh.tools.layout.Views"
+             xmlns:local="clr-namespace:OpenKh.Tools.LayoutViewer.Views"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="400">
     <StackPanel>
