+++ OpenKh.Tools.LayoutViewer/Views/MainWindow.xaml
@@ -1,11 +1,11 @@
-﻿<Window x:Class="kh.tools.layout.Views.MainWindow"
+﻿<Window x:Class="OpenKh.Tools.LayoutViewer.Views.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:kh.tools.layout.Views"
+        xmlns:local="clr-namespace:OpenKh.Tools.LayoutViewer.Views"
         xmlns:controls="clr-namespace:kh.tools.common.Controls;assembly=kh.tools.common"
-        xmlns:localcontrols="clr-namespace:kh.tools.layout.Controls"
+        xmlns:localcontrols="clr-namespace:OpenKh.Tools.LayoutViewer.Controls"
         mc:Ignorable="d"
         Title="{Binding Title}" Height="650" Width="800">
     <Grid>
