+++ OpenKh.Tools.LayoutViewer/Views/MainWindow.xaml.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using OpenKh.Tools.LayoutViewer.ViewModels;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -13,16 +14,17 @@ using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
 
-namespace kh.tools.layout.Views
+namespace OpenKh.Tools.LayoutViewer.Views
 {
     /// <summary>
-    /// Interaction logic for FrameView.xaml
+    /// Interaction logic for MainWindow.xaml
     /// </summary>
-    public partial class FrameView : UserControl
+    public partial class MainWindow : Window
     {
-        public FrameView()
+        public MainWindow()
         {
             InitializeComponent();
+            DataContext = new MainViewModel();
         }
     }
 }
